+++ freq/static/freq/js/map.js
@@ -262,22 +262,6 @@ function loadTimeseries(event) {
   //}
 }
 
-function spinnerShow(){
-  var map = $('#map');
-  var height = map.height() * -0.55;
-  var width = map.width() * 0.49 + "px";
-  $('#wait-spinner')
-    .css('margin-top', height + "px")
-    .css('margin-bottom', -1 * height + 10 + "px")
-    .css('margin-left', width)
-    .removeClass('hidden');
-}
-
-function spinnerClear(){
-  $('#wait-spinner').addClass('hidden');
-}
-
-
 function resetInterpolation(event){
   if (event === undefined){
       event = window.event;